Extract notes feature into separate modules
