+++ .idea/gradle.xml
@@ -14,6 +14,10 @@
             <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/core/designsystem" />
             <option value="$PROJECT_DIR$/core/navigation" />
+            <option value="$PROJECT_DIR$/feature" />
+            <option value="$PROJECT_DIR$/feature/notes" />
+            <option value="$PROJECT_DIR$/feature/notes/api" />
+            <option value="$PROJECT_DIR$/feature/notes/impl" />
           </set>
         </option>
       </GradleProjectSettings>
