+++ app/build.gradle.kts
@@ -3,7 +3,6 @@ plugins {
     alias(libs.plugins.dependency.analysis)
     alias(libs.plugins.hilt)
     alias(libs.plugins.kotlin.compose)
-    alias(libs.plugins.kotlin.serialization)
     alias(libs.plugins.ksp)
     alias(libs.plugins.room)
 }
@@ -73,8 +72,7 @@ android {
             // Enable code shrinking and obfuscation
             isMinifyEnabled = true
             proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
+                getDefaultProguardFile("proguard-android-optimize.txt")
             )
 
             // Remove unused resources
@@ -108,69 +106,36 @@ android {
 dependencies {
     implementation(projects.core.designsystem)
     implementation(projects.core.navigation)
+    implementation(projects.feature.notes.api)
+    implementation(projects.feature.notes.impl)
 
-    // Android core
+    implementation(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.activity)
     implementation(libs.androidx.activity.compose)
     implementation(libs.androidx.annotation)
-    implementation(libs.androidx.core)
-    implementation(libs.androidx.fragment)
-
-    // UI
-    implementation(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.compose.animation)
-    implementation(libs.androidx.compose.animation.core)
     implementation(libs.androidx.compose.foundation)
     implementation(libs.androidx.compose.foundation.layout)
     implementation(libs.androidx.compose.material3)
     implementation(libs.androidx.compose.runtime)
     implementation(libs.androidx.compose.runtime.annotation)
     implementation(libs.androidx.compose.ui)
-    implementation(libs.androidx.compose.ui.geometry)
     implementation(libs.androidx.compose.ui.graphics)
-    implementation(libs.androidx.compose.ui.text)
-    implementation(libs.androidx.compose.ui.unit)
-
-    // Navigation
+    implementation(libs.androidx.fragment)
+    implementation(libs.androidx.lifecycle.viewmodel)
+    implementation(libs.androidx.lifecycle.viewmodel.savedstate)
     implementation(libs.androidx.navigation3.ui)
     implementation(libs.androidx.navigation3.runtime)
-
-    // Concurrency
-    implementation(libs.kotlinx.coroutines.core)
-
-    // Serialization
-    implementation(libs.kotlinx.serialization.core)
-
-    // Architecture
-    implementation(libs.androidx.lifecycle.common)
-    implementation(libs.androidx.lifecycle.runtime.compose)
-    implementation(libs.androidx.lifecycle.viewmodel)
-    implementation(libs.androidx.lifecycle.viewmodel.compose)
-    implementation(libs.androidx.lifecycle.viewmodel.savedstate)
-
-    // Dependency injection
-    implementation(libs.dagger)
-    implementation(libs.dagger.hilt.core)
-    implementation(libs.hilt.android)
-    implementation(libs.hilt.lifecycle.viewmodel.compose)
-    implementation(libs.javax.inject)
-    ksp(libs.hilt.compiler)
-
-    // Data persistence
     implementation(libs.androidx.room.common)
     implementation(libs.androidx.room.runtime)
     implementation(libs.androidx.sqlite)
+    implementation(libs.dagger)
+    implementation(libs.dagger.hilt.core)
+    implementation(libs.hilt.android)
+    implementation(libs.javax.inject)
+    implementation(libs.kotlinx.coroutines.core)
+
     ksp(libs.androidx.room.compiler)
-
-    // Color picker
-    implementation(libs.colorpicker.compose)
-
-    // Metadata
+    ksp(libs.hilt.compiler)
     ksp(libs.kotlin.metadata.jvm)
-
-    // Testing
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.junit)
-    androidTestImplementation(libs.androidx.test.monitor)
-    androidTestImplementation(libs.junit)
 }
