--- app/proguard-rules.pro
@@ -1,1 +1,0 @@
-# ProGuard rules
