--- app/src/androidTest/kotlin/com/oussamateyib/thoth/ExampleInstrumentedTest.kt
@@ -1,17 +1,0 @@
-package com.oussamateyib.thoth
-
-import androidx.test.platform.app.InstrumentationRegistry
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.Assert.*
-
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.oussamateyib.thoth", appContext.packageName)
-    }
-}
