--- app/src/test/kotlin/com/oussamateyib/thoth/ExampleUnitTest.kt
@@ -1,11 +1,0 @@
-package com.oussamateyib.thoth
-
-import org.junit.Test
-import org.junit.Assert.*
-
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
