+++ feature/notes/impl/src/main/kotlin/com/oussamateyib/thoth/feature/notes/impl/presentation/editor/components/ColorPicker.kt
@@ -1,0 +1,40 @@
+package com.oussamateyib.thoth.feature.notes.impl.presentation.editor.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.unit.dp
+import com.github.skydoves.colorpicker.compose.HsvColorPicker
+import com.github.skydoves.colorpicker.compose.rememberColorPickerController
+
+@Composable
+internal fun ColorPicker(
+    selectedColor: Int,
+    onColorChange: (Int) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val controller = rememberColorPickerController()
+
+    Column(
+        modifier = modifier.fillMaxWidth(),
+        verticalArrangement = Arrangement.spacedBy(20.dp)
+    ) {
+        HsvColorPicker(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(250.dp),
+            controller = controller,
+            initialColor = Color(selectedColor),
+            onColorChanged = { envelope ->
+                if (envelope.fromUser) {
+                    onColorChange(envelope.color.toArgb())
+                }
+            }
+        )
+    }
+}
