+++ feature/notes/impl/src/main/kotlin/com/oussamateyib/thoth/feature/notes/impl/presentation/editor/components/TransparentHintTextField.kt
@@ -1,0 +1,46 @@
+package com.oussamateyib.thoth.feature.notes.impl.presentation.editor.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusState
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.text.TextStyle
+import com.oussamateyib.thoth.core.designsystem.theme.DarkGray
+
+@Composable
+internal fun TransparentHintTextField(
+    text: String,
+    hint: String,
+    modifier: Modifier = Modifier,
+    isHintVisible: Boolean = true,
+    onValueChange: (String) -> Unit,
+    textStyle: TextStyle,
+    singleLine: Boolean = false,
+    onFocusChange: (FocusState) -> Unit
+) {
+    Box(
+        modifier = modifier
+    ) {
+        BasicTextField(
+            value = text,
+            onValueChange = onValueChange,
+            singleLine = singleLine,
+            textStyle = textStyle,
+            modifier = Modifier
+                .fillMaxWidth()
+                .onFocusChanged {
+                    onFocusChange(it)
+                }
+        )
+        if (isHintVisible) {
+            Text(
+                text = hint,
+                color = DarkGray
+            )
+        }
+    }
+}
