+++ feature/notes/impl/src/main/kotlin/com/oussamateyib/thoth/feature/notes/impl/presentation/list/components/NoteItem.kt
@@ -1,0 +1,101 @@
+package com.oussamateyib.thoth.feature.notes.impl.presentation.list.components
+
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.CornerRadius
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Path
+import androidx.compose.ui.graphics.drawscope.clipPath
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import androidx.core.graphics.ColorUtils
+import com.oussamateyib.thoth.feature.notes.impl.domain.model.Note
+import com.oussamateyib.thoth.feature.notes.impl.R
+
+@Composable
+internal fun NoteItem(
+    note: Note,
+    modifier: Modifier = Modifier,
+    cornerRadius: Dp = 10.dp,
+    cutCornerSize: Dp = 30.dp,
+    onDeleteClick: () -> Unit
+) {
+    Box(
+        modifier = modifier
+    ) {
+        Canvas(
+            modifier = Modifier.matchParentSize() // Avoid influencing the Box's size
+        ) {
+            val clipPath = Path().apply {
+                lineTo(size.width - cutCornerSize.toPx(), 0f)
+                lineTo(size.width, cutCornerSize.toPx())
+                lineTo(size.width, size.height)
+                lineTo(0f, size.height)
+                close()
+            }
+
+            clipPath(clipPath) {
+                drawRoundRect(
+                    color = Color(note.color),
+                    cornerRadius = CornerRadius(cornerRadius.toPx())
+                )
+                drawRoundRect(
+                    color = Color(
+                        ColorUtils.blendARGB(note.color, android.graphics.Color.BLACK, 0.3f)
+                    ),
+                    topLeft = Offset(size.width - cutCornerSize.toPx(), -100f),
+                    size = Size(cutCornerSize.toPx() + 100f, cutCornerSize.toPx() + 100f),
+                    cornerRadius = CornerRadius(cornerRadius.toPx())
+                )
+            }
+        }
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp)
+                .padding(end = 32.dp)
+        ) {
+            Text(
+                text = note.title,
+                style = MaterialTheme.typography.titleLarge,
+                color = MaterialTheme.colorScheme.onSurface,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis
+            )
+            Spacer(modifier = Modifier.height(8.dp))
+            Text(
+                text = note.content,
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onSurface,
+                maxLines = 10,
+                overflow = TextOverflow.Ellipsis
+            )
+        }
+        IconButton(
+            onClick = onDeleteClick,
+            modifier = Modifier.align(Alignment.BottomEnd)
+        ) {
+            Icon(
+                painter = painterResource(R.drawable.delete),
+                contentDescription = stringResource(R.string.delete_note)
+            )
+        }
+    }
+}
