+++ feature/notes/impl/src/main/kotlin/com/oussamateyib/thoth/feature/notes/impl/presentation/list/components/OrderSection.kt
@@ -1,0 +1,62 @@
+package com.oussamateyib.thoth.feature.notes.impl.presentation.list.components
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.width
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.oussamateyib.thoth.feature.notes.impl.domain.util.NoteOrder
+import com.oussamateyib.thoth.feature.notes.impl.domain.util.OrderType
+import com.oussamateyib.thoth.feature.notes.impl.R
+
+@Composable
+internal fun OrderSection(
+    modifier: Modifier = Modifier,
+    noteOrder: NoteOrder = NoteOrder.Date(OrderType.Descending),
+    onOrderChange: (NoteOrder) -> Unit
+) {
+    Column(
+        modifier = modifier
+    ) {
+        Row(
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            DefaultRadioButton(
+                text = stringResource(R.string.color),
+                selected = noteOrder is NoteOrder.Color,
+                onSelect = { onOrderChange(NoteOrder.Color(noteOrder.orderType)) }
+            )
+            Spacer(modifier = Modifier.width(8.dp))
+            DefaultRadioButton(
+                text = stringResource(R.string.date),
+                selected = noteOrder is NoteOrder.Date,
+                onSelect = { onOrderChange(NoteOrder.Date(noteOrder.orderType)) }
+            )
+            Spacer(modifier = Modifier.width(8.dp))
+            DefaultRadioButton(
+                text = stringResource(R.string.title),
+                selected = noteOrder is NoteOrder.Title,
+                onSelect = { onOrderChange(NoteOrder.Title(noteOrder.orderType)) }
+            )
+        }
+        Row(
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            DefaultRadioButton(
+                text = stringResource(R.string.ascending),
+                selected = noteOrder.orderType is OrderType.Ascending,
+                onSelect = { onOrderChange(noteOrder.copy(OrderType.Ascending)) }
+            )
+            Spacer(modifier = Modifier.width(8.dp))
+            DefaultRadioButton(
+                text = stringResource(R.string.descending),
+                selected = noteOrder.orderType is OrderType.Descending,
+                onSelect = { onOrderChange(noteOrder.copy(OrderType.Descending)) }
+            )
+        }
+    }
+}
