+++ feature/notes/impl/src/main/res/drawable/add.xml
