+++ feature/notes/impl/src/main/res/drawable/arrow_back.xml
