+++ feature/notes/impl/src/main/res/drawable/delete.xml
