+++ feature/notes/impl/src/main/res/drawable/dropper_eye.xml
