+++ gradle/libs.versions.toml
@@ -11,16 +11,13 @@ fragment = "1.8.9"
 hilt = "2.59.2"
 hiltCompose = "1.3.0"
 inject = "1"
-junit = "4.13.2"
-junitAndroidExt = "1.3.0"
 kotlin = "2.3.21"
 ksp = "2.3.9"
 lifecycle = "2.10.0"
-monitor = "1.8.0"
 navigation3 = "1.1.2"
 room = "2.8.4"
-serialization = "1.11.0"
 sqlite = "2.6.2"
+serialization = "1.11.0"
 
 [libraries]
 androidx-activity = { group = "androidx.activity", name = "activity", version.ref = "activity" }
@@ -42,7 +39,6 @@ androidx-compose-ui-text = { group = "androidx.compose.ui", name = "ui-text" }
 androidx-compose-ui-unit = { group = "androidx.compose.ui", name = "ui-unit" }
 androidx-core = { group = "androidx.core", name = "core", version.ref = "core" }
 androidx-fragment = { group = "androidx.fragment", name = "fragment", version.ref = "fragment" }
-androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitAndroidExt" }
 androidx-lifecycle-common = { group = "androidx.lifecycle", name = "lifecycle-common", version.ref = "lifecycle" }
 androidx-lifecycle-runtime-compose = { group = "androidx.lifecycle", name = "lifecycle-runtime-compose", version.ref = "lifecycle" }
 androidx-lifecycle-viewmodel = { group = "androidx.lifecycle", name = "lifecycle-viewmodel", version.ref = "lifecycle" }
@@ -55,7 +51,6 @@ androidx-room-common = { group = "androidx.room", name = "room-common", version.
 androidx-room-compiler = { group = "androidx.room", name = "room-compiler", version.ref = "room" }
 androidx-room-runtime = { group = "androidx.room", name = "room-runtime-android", version.ref = "room" }
 androidx-sqlite = { group = "androidx.sqlite", name = "sqlite", version.ref = "sqlite" }
-androidx-test-monitor = { group = "androidx.test", name = "monitor", version.ref = "monitor" }
 colorpicker-compose = { group = "com.github.skydoves", name = "colorpicker-compose", version.ref = "colorpicker" }
 dagger = { group = "com.google.dagger", name = "dagger", version.ref = "hilt" }
 dagger-hilt-core = { group = "com.google.dagger", name = "hilt-core", version.ref = "hilt" }
@@ -63,7 +58,6 @@ hilt-android = { group = "com.google.dagger", name = "hilt-android", version.ref
 hilt-compiler = { group = "com.google.dagger", name = "hilt-compiler", version.ref = "hilt" }
 hilt-lifecycle-viewmodel-compose = { group = "androidx.hilt", name = "hilt-lifecycle-viewmodel-compose", version.ref = "hiltCompose" }
 javax-inject = { group = "javax.inject", name = "javax.inject", version.ref = "inject" }
-junit = { group = "junit", name = "junit", version.ref = "junit" }
 kotlin-metadata-jvm = { group = "org.jetbrains.kotlin", name = "kotlin-metadata-jvm", version.ref = "kotlin" }
 kotlinx-coroutines-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-core", version.ref = "coroutines" }
 kotlinx-serialization-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-core", version.ref = "serialization" }
